+++ types/global.d.ts
@@ -1,3 +1,5 @@
+/// <reference types="@emotion/react/types/css-prop" />
+
 export {}
 
 declare global {
